pose/physio_ai/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
Pillow==10.1.0
sqlparse==0.4.4
asgiref==3.7.1
pytz==2024.1
numpy==1.24.3
requests==2.31.0
mediapipe==0.10.14
